Refactor reminder into slack.go
